# docs/cdnAdmin/cdnAdmin_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "cdnAdmin的swagger文档",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "petstore.swagger.io:8080",
    "basePath": "/v2",
    "paths": {
        "/testapi/get-string-by-int/{some_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "登录接口",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AuthReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.AuthRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.AuthReq": {
            "type": "object",
            "properties": {
                "nick": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "api.AuthRes": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastIP": {
                    "type": "string"
                },
                "lastTime": {
                    "type": "integer"
                },
                "nick": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                },
                "state": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "integer"
                }
            }
        }
    }
}